requirements.txt
scrapy
scrapyd-client
scrapyd
requests
flask
plotly
pandas
flask-pymongo
flask-script